phase-5-cloud/backend/requirements.txt
fastapi>=0.104
uvicorn[standard]>=0.24
sqlmodel>=0.0.14
alembic>=1.13
psycopg2-binary>=2.9
python-dotenv>=1.0
httpx>=0.25
pytest>=7.4
pytest-asyncio>=0.23
python-dateutil>=2.8
